Compute blueprint from rounded extremes
Instead of using exact extremes and rounding the resulting size. This becomes
apparent with curved rails, which have an innate width of slightly more than 6
and height slightly more than 2, but the blueprint size snapping ends up
computing the width and height of a lone curved rail as 8x4 because each
extreme coordinate is pulled out toward the nearest even number.